updated extension class and theming
